lib: refactor tree building to do two phase assembly
With this commit we refactor the tree building to a two phase approach,
where the first phase is the assembly of the tree structure, and the
second phase is the materialization of the tree transactions. This
allows for easier extension of the tree building process for assets
later on while enabling reuse of the existing tree structure assembly
logic.