assets: comprehensive asset ARK tree test
This commit adds an end to end test in which we build an asset ARK tree
of three participants, unroll the full tree on-chain and finally sweep
back the leaves to the original participants. The purpose of this test
is to demonstrate that virtual tree building mechancs, virtual tx
chaining, proof materialization and sweeping all work together so we can
use it as a building block/example when we fully integrate assets into
our ARK implementation.