baselib/actor: add abstract Mailbox interface
In this commit, we introduce the Mailbox interface to abstract the actor's
message queue implementation. This abstraction, ported from lnd PR 10142,
allows different mailbox strategies to be plugged in without changing the
actor implementation.

This interface can be used to implement features such as: a persistent
mailbox, back pressure management, etc, etc.