lib/tree: update tree extractors for multi-path extraction
Update the ExtractPathForIndices and ExtractPathForCoSigners methods to
take a list of targets instead of just a single leaf target. These can
then be used to extract sub-trees with more than one leaf. This is
useful for a case where a client, for example, has requested more than
one vtxo in a round. With these helpers, we can still just send a single
tree with all the info required for the client to get paths to all its
requested vtxos.